Add a tip on using dental floss to cut out cinnamon rolls
This commit simply adds a useful tip to cut up the rolled up dough.
Often using a knife can pinch the dough whereas using dental floss and
provides for a nice clean cut.

Typically on a floured surface you can slide the 1.5' length of floss
under the rolled up dough and wrap around and pull together to cut the
dough into rolls.

This is best done to first cut the roll in half then keep halving it
into smaller rolls.